backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2
xgboost==2.1.3
lightgbm==4.5.0
python-multipart==0.0.12
requests==2.32.3
authlib==1.3.2
cryptography==43.0.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.9

# Scheduling and timezone support
apscheduler==3.10.4
pytz==2024.1

# Supabase integration
supabase==2.3.4
postgrest==0.16.0
storage3==0.7.4
